fix(web): extract ContactForm client component and address review feedback
- Move mailto form into apps/web/src/app/ContactForm.tsx so page.tsx
  is a Server Component (no 'use client', useState, or useCallback)
- Soften privacy and status copy: replace 'success' with non-final
  'Trying to open your email app...' and clarify that drafts are
  handled by the browser or local mail app, not stored on our backend
- Accept YouTube URLs without protocol by prepending https:// before
  validation
- Reduce message maxLength from 2000 to 1000 to keep mailto URLs safe
- Update stale layout comment that referenced a Google Fonts <head>
  link; fonts come from the system stack via tailwind.config
- Read CONTACT_EMAIL from NEXT_PUBLIC_CONTACT_EMAIL with fallback to
  the live default
