+++ apps/web/src/app/page.tsx
@@ -1,7 +1,5 @@
-'use client';
-
 import Link from 'next/link';
-import { useCallback, useState } from 'react';
+import ContactForm from './ContactForm';
 
 /* ═══════════════════════════════════════════
    UVAI — Grounded One-Page Positioning Site
@@ -122,86 +120,14 @@ const TEMPLATE_CARDS = [
   },
 ];
 
-const USE_CASES = [
-  'Engineering workflow',
-  'Content workflow',
-  'Research workflow',
-  'Business operations workflow',
-  'Other video-to-action workflow',
-];
-
-const CONTACT_EMAIL = 'viralnowsales@gmail.com';
+const DEFAULT_CONTACT_EMAIL = 'viralnowsales@gmail.com';
+const CONTACT_EMAIL =
+  process.env.NEXT_PUBLIC_CONTACT_EMAIL && process.env.NEXT_PUBLIC_CONTACT_EMAIL.trim().length > 0
+    ? process.env.NEXT_PUBLIC_CONTACT_EMAIL
+    : DEFAULT_CONTACT_EMAIL;
 const REPO_URL = 'https://github.com/groupthinking/EventRelay';
 
-function isEmail(value: string) {
-  return /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(value);
-}
-
-function isYouTube(value: string) {
-  if (!value) return true;
-  try {
-    const u = new URL(value);
-    return ['youtube.com', 'www.youtube.com', 'youtu.be', 'm.youtube.com'].includes(u.hostname);
-  } catch {
-    return false;
-  }
-}
-
-type FormStatus = { kind: 'idle' | 'success' | 'error'; text: string };
-
 export default function Home() {
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [useCase, setUseCase] = useState('');
-  const [videoUrl, setVideoUrl] = useState('');
-  const [message, setMessage] = useState('');
-  const [status, setStatus] = useState<FormStatus>({ kind: 'idle', text: '' });
-
-  const handleSubmit = useCallback(
-    (e: React.FormEvent) => {
-      e.preventDefault();
-      const n = name.trim();
-      const em = email.trim();
-      const uc = useCase.trim();
-      const v = videoUrl.trim();
-      const msg = message.trim();
-
-      if (!n || !em || !uc || !msg) {
-        setStatus({ kind: 'error', text: 'Please fill out name, email, use case, and the short note.' });
-        return;
-      }
-      if (n.length > 100 || msg.length > 2000) {
-        setStatus({ kind: 'error', text: 'Keep the name under 100 characters and the note under 2,000 characters.' });
-        return;
-      }
-      if (!isEmail(em)) {
-        setStatus({ kind: 'error', text: 'Please enter a valid email address.' });
-        return;
-      }
-      if (!isYouTube(v)) {
-        setStatus({ kind: 'error', text: 'If you include a sample video, use a YouTube URL.' });
-        return;
-      }
-
-      const subject = encodeURIComponent('UVAI inbound: ' + uc);
-      const body = encodeURIComponent(
-        [
-          'Name: ' + n,
-          'Email: ' + em,
-          'Use case: ' + uc,
-          v ? 'Sample video: ' + v : 'Sample video: not provided',
-          '',
-          'Note:',
-          msg,
-        ].join('\n'),
-      );
-
-      setStatus({ kind: 'success', text: 'Opening your email app with the note prefilled.' });
-      window.location.href = `mailto:${CONTACT_EMAIL}?subject=${subject}&body=${body}`;
-    },
-    [name, email, useCase, videoUrl, message],
-  );
-
   return (
     <div className="min-h-screen text-white" style={{ background: BG }}>
       {/* ─── NAV ─── */}
@@ -647,12 +573,14 @@ export default function Home() {
                 Send the video workflow you want automated.
               </h2>
               <p className="text-base leading-relaxed mb-4" style={{ color: MUTED }}>
-                Share the type of video, the output you need, and where the result should go. The
-                form opens a prefilled email so there is no hidden backend or fake submission state.
+                Share the type of video, the output you need, and where the result should go.
+                Submitting hands the draft to your browser or local mail app, which fills in a new
+                message you can review and send.
               </p>
               <p className="text-sm leading-relaxed mb-7" style={{ color: FAINT }}>
-                Privacy note: this page does not store form data. Submitting opens your own email
-                client and sends the note directly to the contact inbox.
+                Privacy note: this page does not submit to our backend. The draft contents are
+                handed to your browser or local email app, which formats them as a new message for
+                you to review and send.
               </p>
               <div className="flex flex-wrap gap-3">
                 <Link
@@ -674,106 +602,7 @@ export default function Home() {
               </div>
             </div>
 
-            <article
-              className="rounded-2xl p-7"
-              style={{ background: 'rgba(19,19,24,0.65)', border: `1px solid ${BORDER}` }}
-            >
-              <h3 className="font-heading text-xl font-bold mb-2" style={{ color: INK }}>
-                Workflow request
-              </h3>
-              <p className="text-sm mb-5" style={{ color: MUTED }}>
-                Four fields. Enough to start a useful conversation.
-              </p>
-              <form onSubmit={handleSubmit} className="grid gap-4" noValidate>
-                <Field label="Name" htmlFor="name">
-                  <input
-                    id="name"
-                    name="name"
-                    autoComplete="name"
-                    placeholder="Your name"
-                    maxLength={100}
-                    required
-                    value={name}
-                    onChange={(e) => setName(e.target.value)}
-                    className="form-input"
-                  />
-                </Field>
-                <Field label="Email" htmlFor="email">
-                  <input
-                    id="email"
-                    name="email"
-                    type="email"
-                    autoComplete="email"
-                    placeholder="you@company.com"
-                    required
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    className="form-input"
-                  />
-                </Field>
-                <Field label="Use case" htmlFor="use_case">
-                  <select
-                    id="use_case"
-                    name="use_case"
-                    required
-                    value={useCase}
-                    onChange={(e) => setUseCase(e.target.value)}
-                    className="form-input"
-                  >
-                    <option value="">Choose one</option>
-                    {USE_CASES.map((opt) => (
-                      <option key={opt}>{opt}</option>
-                    ))}
-                  </select>
-                </Field>
-                <Field label="Sample YouTube URL" htmlFor="video_url">
-                  <input
-                    id="video_url"
-                    name="video_url"
-                    type="url"
-                    placeholder="https://youtube.com/watch?v=..."
-                    value={videoUrl}
-                    onChange={(e) => setVideoUrl(e.target.value)}
-                    className="form-input"
-                  />
-                </Field>
-                <Field label="What should the output be?" htmlFor="message">
-                  <textarea
-                    id="message"
-                    name="message"
-                    placeholder="Example: turn product demo videos into API docs and tickets."
-                    maxLength={2000}
-                    required
-                    rows={4}
-                    value={message}
-                    onChange={(e) => setMessage(e.target.value)}
-                    className="form-input"
-                  />
-                </Field>
-                <p
-                  role="status"
-                  aria-live="polite"
-                  className="text-sm min-h-[1.25rem]"
-                  style={{
-                    color:
-                      status.kind === 'success'
-                        ? TEAL
-                        : status.kind === 'error'
-                          ? '#ff8a8a'
-                          : FAINT,
-                  }}
-                >
-                  {status.text}
-                </p>
-                <button
-                  type="submit"
-                  className="px-6 py-3 rounded-lg font-bold text-sm transition-all duration-300 active:scale-95"
-                  style={{ background: `linear-gradient(135deg, ${TEAL}, ${TEAL_DEEP})`, color: '#002b26' }}
-                >
-                  Send request
-                </button>
-              </form>
-            </article>
+            <ContactForm />
           </div>
         </section>
       </main>
@@ -806,62 +635,6 @@ export default function Home() {
           </div>
         </div>
       </footer>
-
-      <style jsx>{`
-        .form-input {
-          width: 100%;
-          padding: 0.7rem 0.9rem;
-          border-radius: 0.55rem;
-          background: rgba(14, 14, 19, 0.7);
-          border: 1px solid ${BORDER};
-          color: ${INK};
-          font-size: 0.9rem;
-          transition: border-color 0.2s, box-shadow 0.2s;
-        }
-        .form-input::placeholder {
-          color: rgba(248, 245, 253, 0.3);
-        }
-        .form-input:focus {
-          outline: none;
-          border-color: ${TEAL};
-          box-shadow: 0 0 0 3px rgba(106, 242, 222, 0.15);
-        }
-        select.form-input {
-          appearance: none;
-          background-image: linear-gradient(45deg, transparent 50%, ${TEAL} 50%),
-            linear-gradient(135deg, ${TEAL} 50%, transparent 50%);
-          background-position: calc(100% - 18px) 50%, calc(100% - 13px) 50%;
-          background-size: 5px 5px, 5px 5px;
-          background-repeat: no-repeat;
-          padding-right: 2.2rem;
-        }
-        textarea.form-input {
-          resize: vertical;
-          min-height: 100px;
-        }
-      `}</style>
     </div>
-  );
-}
-
-function Field({
-  label,
-  htmlFor,
-  children,
-}: {
-  label: string;
-  htmlFor: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <label htmlFor={htmlFor} className="grid gap-1.5">
-      <span
-        className="text-[11px] font-bold uppercase tracking-widest"
-        style={{ color: 'rgba(248,245,253,0.65)' }}
-      >
-        {label}
-      </span>
-      {children}
-    </label>
   );
 }
